@@ -1,5 +1,7 @@
 {
+    "$schema": "https://getcomposer.org/schema.json",
     "name": "statamic/statamic",
+    "type": "project",
     "description": "Statamic",
     "keywords": [
         "statamic",
@@ -7,39 +9,28 @@
         "flat file",
         "laravel"
     ],
-    "type": "project",
     "require": {
         "php": "^8.2",
         "eminos/statamic-tabs": "^1.2",
-        "guzzlehttp/guzzle": "^7.9",
-        "jonassiewertsen/statamic-livewire": "^3.8",
-        "laravel/framework": "^11.24",
-        "laravel/sanctum": "^4.0",
-        "laravel/tinker": "^2.9",
-        "marcorieser/statamic-instagram": "^2.0",
-        "nesbot/carbon": "^2.72",
-        "simonridley/tracking-code-manager": "^1.0",
-        "spatie/browsershot": "^4.3",
-        "statamic/cms": "^5.25",
-        "studio1902/statamic-peak-browser-appearance": "^3.5",
-        "studio1902/statamic-peak-commands": "^8.8",
-        "studio1902/statamic-peak-seo": "^8.16",
-        "studio1902/statamic-peak-tools": "^7.1.0",
+        "laravel/framework": "^12.0",
+        "laravel/tinker": "^2.10.1",
+        "statamic/cms": "^5.0",
+        "studio1902/statamic-peak-browser-appearance": "^3.4",
+        "studio1902/statamic-peak-commands": "^8.0",
+        "studio1902/statamic-peak-seo": "^8.19",
+        "studio1902/statamic-peak-tools": "^7.0",
         "xndbogdan/statamic-bard-text-color": "^5.1"
     },
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": "dist",
-        "sort-packages": true,
-        "allow-plugins": {
-            "composer/package-versions-deprecated": true,
-            "pixelfear/composer-dist-plugin": true
-        }
-    },
-    "extra": {
-        "laravel": {
-            "dont-discover": []
-        }
+    "require-dev": {
+        "barryvdh/laravel-debugbar": "^3.8.1",
+        "fakerphp/faker": "^1.23",
+        "laravel/pail": "^1.2.2",
+        "laravel/pint": "^1.13",
+        "laravel/sail": "^1.41",
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.6",
+        "phpunit/phpunit": "^11.5.3",
+        "spatie/laravel-ignition": "^2.9.1"
     },
     "autoload": {
         "psr-4": {
@@ -53,8 +44,6 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
     "scripts": {
         "pre-update-cmd": [
             "Statamic\\Console\\Composer\\Scripts::preUpdateCmd"
@@ -68,19 +57,32 @@
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "@php artisan key:generate --ansi"
+            "@php artisan key:generate --ansi",
+            "@php -r \"file_exists('database/database.sqlite') || touch('database/database.sqlite');\""
         ],
         "post-update-cmd": [
             "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
+        ],
+        "dev": [
+            "Composer\\Config::disableProcessTimeout",
+            "npx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan pail --timeout=0\" \"npm run dev\" --names=server,queue,logs,vite"
         ]
     },
-    "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.14",
-        "fakerphp/faker": "^1.24",
-        "laravel/sail": "^1.41",
-        "mockery/mockery": "^1.6",
-        "nunomaduro/collision": "^8.4",
-        "phpunit/phpunit": "^11.3",
-        "spatie/laravel-ignition": "^2.8"
-    }
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true,
+            "pixelfear/composer-dist-plugin": true
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
