@@ -3,6 +3,7 @@
 namespace App\Modifiers;
 
 use Statamic\Modifiers\Modifier;
+use Illuminate\Support\Arr; // Laravel 6.x and up
 
 class IsYoutube extends Modifier
 {
@@ -26,10 +27,19 @@ class IsYoutube extends Modifier
             $is_youtube = true;
         }
 
-        if ($param = array_get($params, 0)) {
+        //$param = array_get($params, 0)
+        $param  = Arr::get($params, 0, null); // Replace array_get with Arr::get
+        if ($param !== null) {
             if( $param == 'id' 
                 && $is_youtube
             ) {
+                if(strpos($value, "shorts") !== false) {
+                    //https://www.youtube.com/shorts/E0SVJLHV4MY
+                    $youtubeid = explode('shorts/', $value);
+                    $youtubeid = $youtubeid[1];
+                    return $youtubeid;
+                }
+
                 $youtubeid = explode('v=', $value);
                 $youtubeid = $youtubeid[1];
                 $youtubeid = explode('?', $youtubeid);
