@@ -2,17 +2,31 @@
 
 namespace App\Providers;
 
+use App\Policies\CustomUserPolicy;
+use Illuminate\Cache\RateLimiting\Limit;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\RateLimiter;
 use Illuminate\Support\ServiceProvider;
+use Statamic\Policies\UserPolicy;
 use Studio1902\PeakSeo\Handlers\ErrorPage;
 
 class AppServiceProvider extends ServiceProvider
 {
     /**
+     * The path to your application's "home" route.
+     *
+     * Typically, users are redirected here after authentication.
+     *
+     * @var string
+     */
+    public const HOME = '/home';
+
+    /**
      * Register any application services.
      */
     public function register(): void
     {
-        //
+        $this->app->bind(UserPolicy::class, CustomUserPolicy::class);
     }
 
     /**
@@ -24,5 +38,14 @@ class AppServiceProvider extends ServiceProvider
         // Statamic::style('app', 'cp');
 
         ErrorPage::handle404AsEntry();
+
+        $this->bootRoute();
+    }
+
+    public function bootRoute(): void
+    {
+        RateLimiter::for('api', function (Request $request) {
+            return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
+        });
     }
 }
