@@ -1,7 +1,7 @@
 <?php
 
 return [
-    
+
     /*
     |--------------------------------------------------------------------------
     | Bard Custom Colors
@@ -19,7 +19,7 @@ return [
     | ]
     |
     */
-    
+
 
     /*
     |--------------------------------------------------------------------------
@@ -44,12 +44,6 @@ return [
 
 
     'default_colors' => [
-        'black'     => '#1D252C',
-        'blue'      => '#0083C1',   
-        'green'     => '#00AD41',
-        'grey'      => '#D9D8D6',
-        'orange'    => '#FC6D3A',
-        'white'     => '#FFFFFF',
         'black'     => '#000',
         'blue'      => '#0071CE',
         'darkblue'  => '#000F9F',
