@@ -9,33 +9,32 @@ return [
     |
     | Here you may specify the default filesystem disk that should be used
     | by the framework. The "local" disk, as well as a variety of cloud
-    | based disks are available to your application. Just store away!
+    | based disks are available to your application for file storage.
     |
     */
 
     'default' => env('FILESYSTEM_DISK', 'local'),
-
-    'cloud' => env('FILESYSTEM_CLOUD', 's3'),
 
     /*
     |--------------------------------------------------------------------------
     | Filesystem Disks
     |--------------------------------------------------------------------------
     |
-    | Here you may configure as many filesystem "disks" as you wish, and you
-    | may even configure multiple disks of the same driver. Defaults have
-    | been set up for each driver as an example of the required values.
+    | Below you may configure as many filesystem disks as necessary, and you
+    | may even configure multiple disks for the same driver. Examples for
+    | most supported storage drivers are configured here for reference.
     |
     | Supported Drivers: "local", "ftp", "sftp", "s3"
     |
     */
 
     'disks' => [
-
         'local' => [
             'driver' => 'local',
-            'root' => storage_path('app'),
+            'root' => storage_path('app/private'),
+            'serve' => true,
             'throw' => false,
+            'report' => false,
         ],
 
         'public' => [
@@ -44,18 +43,16 @@ return [
             'url' => env('APP_URL').'/storage',
             'visibility' => 'public',
             'throw' => false,
+            'report' => false,
         ],
 
-        's3' => [
-            'driver' => 's3',
-            'key' => env('AWS_ACCESS_KEY_ID'),
-            'secret' => env('AWS_SECRET_ACCESS_KEY'),
-            'region' => env('AWS_DEFAULT_REGION'),
-            'bucket' => env('AWS_BUCKET'),
-            'url' => env('AWS_URL'),
-            'endpoint' => env('AWS_ENDPOINT'),
-            'use_path_style_endpoint' => env('AWS_USE_PATH_STYLE_ENDPOINT', false),
+        'assets' => [
+            'driver' => 'local',
+            'root' => public_path('assets'),
+            'url' => '/assets',
+            'visibility' => 'public',
             'throw' => false,
+            'report' => false,
         ],
 
         'images' => [
@@ -63,6 +60,8 @@ return [
             'root' => public_path('images'),
             'url' => '/images',
             'visibility' => 'public',
+            'throw' => false,
+            'report' => false,
         ],
 
         'favicons' => [
@@ -70,6 +69,8 @@ return [
             'root' => public_path('favicons'),
             'url' => '/favicons',
             'visibility' => 'public',
+            'throw' => false,
+            'report' => false,
         ],
 
         'files' => [
@@ -77,6 +78,8 @@ return [
             'root' => public_path('files'),
             'url' => '/files',
             'visibility' => 'public',
+            'throw' => false,
+            'report' => false,
         ],
 
         'social_images' => [
@@ -84,8 +87,9 @@ return [
             'root' => public_path('social_images'),
             'url' => '/social_images',
             'visibility' => 'public',
+            'throw' => false,
+            'report' => false,
         ],
-
     ],
 
     /*
