@@ -17,6 +17,21 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Authentication
+    |--------------------------------------------------------------------------
+    |
+    | Whether the Control Panel's authentication pages should be enabled,
+    | and where users should be redirected in order to authenticate.
+    |
+    */
+
+    'auth' => [
+        'enabled' => true,
+        'redirect_to' => null,
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
     | Start Page
     |--------------------------------------------------------------------------
     |
@@ -41,25 +56,13 @@ return [
         [
             'type' => 'images_missing_alt',
             'container' => 'images',
-            'filetypes' => ['jpg', 'jpeg', 'png', 'gif', 'webp', 'avif', 'bmp', 'tiff', 'svg'],
             'limit' => 5,
-            'expiry' => 0,
             'width' => 50,
         ],
         [
             'type' => 'collection',
             'collection' => 'pages',
             'width' => 50,
-        ],
-        [
-            'type' => 'collection',
-            'collection' => 'news',
-            'width' => 50
-        ],
-        [
-            'type' => 'collection',
-            'collection' => 'projects',
-            'width' => 50
         ],
         [
             'type' => 'form',
@@ -152,4 +155,16 @@ return [
 
     'custom_css_url' => env('STATAMIC_CUSTOM_CSS_URL', null),
 
+    /*
+    |--------------------------------------------------------------------------
+    | Thumbnails
+    |--------------------------------------------------------------------------
+    |
+    | Here you may define additional CP asset thumbnail presets.
+    |
+    */
+
+    'thumbnail_presets' => [
+        // 'medium' => 800,
+    ],
 ];
