@@ -2,7 +2,7 @@
 
 return [
 
-    'pro' => true,
+    'pro' => env('STATAMIC_PRO_ENABLED', false),
 
     'addons' => [
         //
