@@ -54,11 +54,14 @@ return [
     */
 
     'exclude' => [
-        '/site.webmanifest',
-        '/sitemap.xml',
-        '/sitemaps.xml',
-        '/feed*', // Exclude all feeds
-        // Add sitemaps for multisites here
+
+        'class' => null,
+
+        'urls' => [
+            '/site.webmanifest',
+            '/sitemap.xml',
+            '/sitemaps.xml',
+        ],
     ],
 
     /*
@@ -95,6 +98,31 @@ return [
 
     'ignore_query_strings' => false,
 
+    'allowed_query_strings' => [
+        //
+    ],
+
+    'disallowed_query_strings' => [
+        //
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Nocache
+    |--------------------------------------------------------------------------
+    |
+    | Here you may define where the nocache data is stored.
+    |
+    | https://statamic.dev/tags/nocache#database
+    |
+    | Supported drivers: "cache", "database"
+    |
+    */
+
+    'nocache' => 'cache',
+
+    'nocache_js_position' => 'body',
+
     /*
     |--------------------------------------------------------------------------
     | Replacers
@@ -115,11 +143,13 @@ return [
     | Warm Queue
     |--------------------------------------------------------------------------
     |
-    | Here you may define the name of the queue that requests will be pushed
-    | onto when warming the static cache using the static:warm command.
+    | Here you may define the queue name and connection
+    | that will be used when warming the static cache.
     |
     */
 
-    'warm_queue' => null,
+    'warm_queue' => env('STATAMIC_STATIC_WARM_QUEUE'),
+
+    'warm_queue_connection' => env('STATAMIC_STATIC_WARM_QUEUE_CONNECTION'),
 
 ];
