@@ -46,7 +46,7 @@ return [
     |
     */
 
-    'avatars' => 'gravatar',
+    'avatars' => 'initials',
 
     /*
     |--------------------------------------------------------------------------
@@ -78,20 +78,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | User Wizard Invitation Email
-    |--------------------------------------------------------------------------
-    |
-    | When creating new users through the wizard in the control panel,
-    | you may choose whether to be able to send an invitation email.
-    | Setting to true will give the user the option. But setting
-    | it to false will disable the invitation option entirely.
-    |
-    */
-
-    'wizard_invitation' => true,
-
-    /*
-    |--------------------------------------------------------------------------
     | Password Brokers
     |--------------------------------------------------------------------------
     |
@@ -102,8 +88,8 @@ return [
     */
 
     'passwords' => [
-        'resets' => 'resets',
-        'activations' => 'activations',
+        'resets' => config('auth.defaults.passwords'),
+        'activations' => config('auth.defaults.passwords'),
     ],
 
     /*
