@@ -9,7 +9,7 @@ return [
     'form_mail_closing' => 'S přátelským pozdravem',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Vážený',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
     'form_mail_submitted' => 'Odesláno',
     'form_mail_title' => 'Kontaktní formulář',
     'form_mail_url' => config('app.url'),
@@ -19,28 +19,31 @@ return [
     'form_success' => 'Děkujeme, vaši zprávu jsme obdrželi a budeme vás co nejdříve kontaktovat.',
 
     // Cookie banner
-    'cookie_accept_all' => 'Přijmout vše',
-    'cookie_accept_selected' => 'Přijmout vybrané',
-    'cookie_ads' => 'Reklamy & sledování',
-    'cookie_ads_explanation' => 'Third party cookies absolutely terrible for privacy, but we don\'t give a sh*t.',
-    'cookie_analytics' => 'Analytické',
-    'cookie_analytics_explanation' => 'Third party cookies, not needed if we cared.',
-    'cookie_explanation' => 'Tento web používá cookies, které vyžadují souhlas.',
-    'cookie_embeds' => 'Video embeds',
-    'cookie_embeds_explanation' => 'Cookie soubory třetích stran používané pro vložená videa, která vás mohou sledovat.',
-    'cookie_embeds_disabled' => 'Chcete-li zobrazit toto video, přijměte soubory cookie pro vložení videa.',
-    'cookie_functional' => 'Funkční',
-    'cookie_functional_explanation' => 'Cookies potřebné pro fungování webu. Tyto nelze vypnout.',
-    'cookie_title' => 'Cookies 🍪',
-    'cookie_learn_more' => 'Zjistit více',
-    'cookie_change_preferences' => 'Změnit cookie preference',
+    'consent_accept_all' => 'Přijmout vše',
+    'consent_accept_selected' => 'Přijmout vybrané',
+    'consent_reject_all' => 'Reject all',
+    'consent_ads' => 'Reklamy & sledování',
+    'consent_ads_explanation' => 'Third party cookies absolutely terrible for privacy, but we don\'t give a sh*t.',
+    'consent_analytics' => 'Analytické',
+    'consent_analytics_explanation' => 'Third party cookies, not needed if we cared.',
+    'consent_explanation' => 'Tento web používá cookies, které vyžadují souhlas.',
+    'consent_embeds' => 'Video embeds',
+    'consent_embeds_explanation' => 'Cookie soubory třetích stran používané pro vložená videa, která vás mohou sledovat.',
+    'consent_embeds_accept' => 'Udělit souhlas se zobrazením tohoto obsahu vloženým třetím stranám.',
+    'consent_embeds_open' => 'Otevřete banner souhlasu s přijímáním vložených souborů.',
+    'consent_functional' => 'Funkční',
+    'consent_functional_explanation' => 'Cookies potřebné pro fungování webu. Tyto nelze vypnout.',
+    'consent_title' => 'Cookies 🍪',
+    'consent_learn_more' => 'Zjistit více',
+    'consent_change_preferences' => 'Změnit cookie preference',
 
     // Footer
     'on' => 'on',
 
     // Navigation
     'close' => 'Zavřít',
-    'main_navigation' => 'Hlavní navigace',
+    'main_navigation_desktop' => 'Hlavní navigace desktop',
+    'main_navigation_mobile' => 'Hlavní navigace mobile',
     'menu' => 'Menu',
     'nav_close' => 'Zavřít navigaci',
     'nav_open' => 'Otevřít navigaci',
@@ -49,6 +52,7 @@ return [
     'skip_to_content' => 'Přejít na obsah',
 
     // Other
+    'home' => 'Homepage',
     'no_results' => 'Momentálně zde nejsou žádné příspěvky',
     'no_script' => 'Tato webová stránka vyžaduje pro správnou funkci některých částí Javascript. Váš uživatelský požitek se může lišit.',
 
@@ -56,15 +60,4 @@ return [
     'next' => 'další',
     'of' => 'z',
     'previous' => 'předchozí',
-
-    // Social images
-    'social_images' => 'Vygenerován jeden obrázek pro soc. sítě.|Vygenerováno :count obrázků pro soc. sítě.',
-    'social_images_queue' => 'Generování jednoho obrázku pro soc. sítě na pozadí.|Generování :count obrázků pro soc. sítě na pozadí.',
-
-    // Widgets
-    'widget_assets_title' => 'Assety bez alternativního textu v :container',
-    'widget_assets_edit' => 'Editovat tento Asset',
-    'widget_assets_explanation' => 'Je důležité přidat alternativní text popisující vaše obrázky. Pomůže to uživatelům, kteří jsou závislí na asistenčních technologiích.',
-    'widget_assets_count' => '{0}|{1}Máte :amount obrázek, který potřebuje vaši pozornost.|[2,4]Máte :amount obrázky, které potřebují vaši pozornost.|[5,*]Máte :amount obrázeků, které potřebují vaši pozornost.',
-    'widget_assets_done' => 'Všechny Assety mají alternativní text.',
 ];
