@@ -3,39 +3,41 @@
 return [
     // Contact form
     'form_honeypot' => 'Fax',
-    'form_mail_body_owner' => 'Der er sendt en kontaktformular.',
+    'form_mail_body_owner' => 'Der er sendt en besked.',
     'form_mail_body_sender' => 'Tak for din besked. Vi kontakter dig hurtigst muligt.',
-    'form_mail_closing' => 'Med venlig hilsens',
+    'form_mail_closing' => 'Med venlig hilsen',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Kære',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
     'form_mail_submitted' => 'Sendt',
     'form_mail_title' => 'Kontaktformular',
     'form_mail_url' => config('app.url'),
     'form_send' => 'Send besked',
-    'form_subject_received' => 'Kontaktformular modtaget',
-    'form_subject_sent' => 'Kontaktformular sendt',
+    'form_subject_received' => 'Besked modtaget',
+    'form_subject_sent' => 'Besked sendt',
     'form_success' => 'Tak, vi har modtaget din besked og kontakter dig hurtigst muligt.',
 
     // Cookie banner
-    'cookie_accept' => 'Acceptere',
-    'cookie_ads' => 'Annoncer og sporing',
-    'cookie_ads_explanation' => 'Tredjepartscookies brugt til reklame.',
-    'cookie_analytics' => 'Analyser',
-    'cookie_analytics_explanation' => 'Tredjepartscookies brugt til analyser.',
-    'cookie_explanation' => 'Denne side bruger cookies, der kræver samtykke.',
-    'cookie_embeds' => 'Video embeds',
-    'cookie_embeds_explanation' => 'Tredjepartscookies, der bruges til indlejrede videoer, der muligvis kan spore dig.',
-    'cookie_embeds_disabled' => 'Accepter videoindlejringscookies for at se denne video.',
-    'cookie_functional' => 'Funktionel',
-    'cookie_functional_explanation' => 'Førstepartscookies er nødvendige for at hjemmesiden kan fungere. Disse kan ikke slås fra.',
-    'cookie_ignore' => 'Ignorere',
-    'cookie_title' => 'Cookies 🍪',
-    'cookie_learn_more' => 'Lær mere',
-    'cookie_reset_consent' => 'Nulstil cookie-samtykke',
+    'consent_accept' => 'Acceptere',
+    'consent_reject_all' => 'Reject all',
+    'consent_ads' => 'Annoncer og sporing',
+    'consent_ads_explanation' => 'Tredjepartscookies brugt til reklame.',
+    'consent_analytics' => 'Analyser',
+    'consent_analytics_explanation' => 'Tredjepartscookies brugt til analyser.',
+    'consent_explanation' => 'Denne side bruger cookies, der kræver samtykke.',
+    'consent_embeds' => 'Video embeds',
+    'consent_embeds_explanation' => 'Tredjepartscookies, der bruges til indlejrede videoer, der muligvis kan spore dig.',
+    'consent_embeds_accept' => 'Giv samtykke til tredjepartsindlejringer for at se dette indhold.',
+    'consent_embeds_open' => 'Åbn samtykkebanneret for at acceptere indlejringer.',
+    'consent_functional' => 'Funktionel',
+    'consent_functional_explanation' => 'Førstepartscookies er nødvendige for at hjemmesiden kan fungere. Disse kan ikke slås fra.',
+    'consent_ignore' => 'Ignorere',
+    'consent_title' => 'Cookies 🍪',
+    'consent_learn_more' => 'Lær mere',
+    'consent_reset_consent' => 'Nulstil cookie-samtykke',
 
     // Footer
-    'on' => 'Aktiv',
+    'on' => 'på',
 
     // Navigation
     'close' => 'Luk',
@@ -48,6 +50,7 @@ return [
     'skip_to_content' => 'Gå til indhold',
 
     // Other
+    'home' => 'Homepage',
     'no_results' => 'Der er i øjeblikket ingen indlæg.',
     'no_script' => 'Denne hjemmeside kræver Javascript for at nogle dele kan fungere korrekt. Din oplevelse kan variere.',
 
@@ -55,15 +58,4 @@ return [
     'next' => 'næste',
     'of' => 'af',
     'previous' => 'forrige',
-
-    // Social images
-    'social_images' => 'Genereret ét socialt billede.|Genereret: tæl sociale billeder.',
-    'social_images_queue' => 'Generering af ét socialt billede i baggrunden.|Generering :count sociale billeder i baggrunden.',
-
-    // Widgets
-    'widget_assets_title' => 'Aktiver uden Alt Text i :container',
-    'widget_assets_edit' => 'Rediger dette aktiv',
-    'widget_assets_explanation' => 'Det er vigtigt at tilføje alt-tekst, der beskriver dine billeder. Dette hjælper brugere, der er afhængige af hjælpeteknologi.',
-    'widget_assets_count' => '{0}Du har :amount billeder, der kræver opmærksomhed.|{1}Du har :amount billeder, der kræver opmærksomhed.|[2,*]Du har mindst :amount billeder, der kræver opmærksomhed.',
-    'widget_assets_done' => 'Alle aktiver har en alternativ tekst.',
 ];
