@@ -9,7 +9,7 @@ return [
     'form_mail_closing' => 'Herzliche Grüße',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Hallo',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
     'form_mail_submitted' => 'gesendet',
     'form_mail_title' => 'Kontaktformular',
     'form_mail_url' => config('app.url'),
@@ -19,28 +19,36 @@ return [
     'form_success' => 'Vielen Dank für Ihre Nachricht. Wir werden uns in Kürze bei Ihnen melden.',
 
     // Cookie banner
-    'cookie_accept_all' => 'Alle akzeptieren',
-    'cookie_accept_selected' => 'Ausgewählte akzeptieren',
-    'cookie_ads' => 'Werbung & Tracking',
-    'cookie_ads_explanation' => 'Marketing-Cookies werden verwendet, um Besuchern auf Webseiten zu folgen und ggf. relevante Werbung anzuzeigen.',
-    'cookie_analytics' => 'Analyse',
-    'cookie_analytics_explanation' => 'Analyse-Cookies helfen Webseiten-Besitzern zu verstehen, wie Besucher mit Webseiten interagieren, indem Informationen anonym gesammelt und gemeldet werden.',
-    'cookie_explanation' => 'Diese Website verwendet Cookies, die Ihre Zustimmung brauchen.',
-    'cookie_embeds' => 'Video Einbettungen',
-    'cookie_embeds_explanation' => 'Cookies von Drittparteien, die für eingebettete Videos verwendet werden.',
-    'cookie_embeds_disabled' => 'Akzeptieren Sie Videoeinbettungs-Cookies, um dieses Video anzusehen.',
-    'cookie_functional' => 'Funktional',
-    'cookie_functional_explanation' => 'Funktionale Cookies sind für die Funktionsfähigkeit der Website notwendig. Diese Cookies Können nicht deaktiviert werden.',
-    'cookie_title' => 'Cookies 🍪',
-    'cookie_learn_more' => 'Details',
-    'cookie_change_preferences' => 'Cookie-Einstellungen bearbeiten',
+    'consent_accept_all' => 'Alle akzeptieren',
+    'consent_accept_selected' => 'Ausgewählte akzeptieren',
+    'consent_reject_all' => 'Alle ablehnen',
+    'consent_ads' => 'Werbung & Tracking',
+    'consent_ads_explanation' => 'Marketing-Cookies werden verwendet, um Besuchern auf Webseiten zu folgen und ggf. relevante Werbung anzuzeigen.',
+    'consent_ad_user_data' => 'Nutzerdaten',
+    'consent_ad_user_data_explanation' => 'Übermittlung von werbebezogenen Nutzerdaten an Google.',
+    'consent_ad_personalization' => 'Personalisierung von Anzeigen',
+    'consent_ad_personalization_explanation' => 'Personalisierte Werbung erhalten.',
+    'consent_analytics' => 'Analyse',
+    'consent_analytics_explanation' => 'Analyse-Cookies helfen Webseiten-Besitzern zu verstehen, wie Besucher mit Webseiten interagieren, indem Informationen anonym gesammelt und gemeldet werden.',
+    'consent_explanation' => 'Diese Website verwendet Cookies, die Ihre Zustimmung brauchen.',
+    'consent_embeds' => 'Video Einbettungen',
+    'consent_embeds_explanation' => 'Cookies von Drittparteien, die für eingebettete Videos verwendet werden.',
+    'consent_embeds_accept' => 'Einbettungen akzeptieren, um diesen Inhalt anzusehen.',
+    'consent_embeds_open' => 'Einwilligungsbanner öffnen, um Einbettungen zu akzeptieren.',
+    'consent_functional' => 'Funktional',
+    'consent_functional_explanation' => 'Funktionale Cookies sind für die Funktionsfähigkeit der Website notwendig. Diese Cookies Können nicht deaktiviert werden.',
+    'consent_title' => 'Cookies 🍪',
+    'consent_learn_more' => 'Details',
+    'consent_change_preferences' => 'Cookie-Einstellungen bearbeiten',
+    'consent_more_info' => 'Mehr Information',
 
     // Footer
     'on' => 'auf',
 
     // Navigation
     'close' => 'Schließen',
-    'main_navigation' => 'Hauptnavigation',
+    'main_navigation_desktop' => 'Hauptnavigation desktop',
+    'main_navigation_mobile' => 'Hauptnavigation mobile',
     'menu' => 'Menü',
     'nav_close' => 'Navigation schließen',
     'nav_open' => 'Navigation öffnen',
@@ -49,6 +57,7 @@ return [
     'skip_to_content' => 'Zum Inhalt springen',
 
     // Other
+    'home' => 'Homepage',
     'no_results' => 'Momentan sind keine Einträge vorhanden.',
     'no_script' => 'Diese Website funktioniert am besten, wenn JavaScript aktiviert ist.',
 
@@ -56,15 +65,4 @@ return [
     'next' => 'weiter',
     'of' => 'von',
     'previous' => 'zurück',
-
-    // Social images
-    'social_images' => 'Ein social image generiert.|:count social images generiert.',
-    'social_images_queue' => 'Generiere ein social image im Hintergrund.|Generiere :count social images im Hintergrund.',
-
-    // Widgets
-    'widget_assets_title' => 'Bilder ohne Alternativtext in :container',
-    'widget_assets_edit' => 'Dieses Bild bearbeiten',
-    'widget_assets_explanation' => 'Es ist wichtig, Bilder mit Alternativtexten zu beschreiben. Dies hilft Benutzer:innen von assistiver Technologie, die Inhalte zu erfassen.',
-    'widget_assets_count' => '{0}|{1} :amount Bild benötigt Aufmerksamkeit.|[2,*]Mindestens :amount Bilder benötigen Aufmerksamkeit.',
-    'widget_assets_done' => 'Alle Bilder verfügen über eine Alternativtext.',
 ];
