@@ -1,19 +1,8 @@
 <?php
 
+declare(strict_types=1);
+
 return [
-
-    /*
-    |--------------------------------------------------------------------------
-    | Pagination Language Lines
-    |--------------------------------------------------------------------------
-    |
-    | The following language lines are used by the paginator library to build
-    | the simple pagination links. You are free to change them to anything
-    | you want to customize your views to better match your application.
-    |
-    */
-
+    'next'     => 'Next &raquo;',
     'previous' => '&laquo; Previous',
-    'next' => 'Next &raquo;',
-
 ];
