@@ -2,14 +2,14 @@
 
 return [
     // Contact form
-    'form_error' => 'There are errors with your form input. Please fix them and try again.',
+    'form_error' => 'There are problems with your form input:',
     'form_honeypot' => 'Fax',
     'form_mail_body_owner' => 'A contact form has been sent.',
     'form_mail_body_sender' => 'Thanks for your message. We will contact you as soon as possible.',
     'form_mail_closing' => 'Kind regards',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Dear',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
     'form_mail_submitted' => 'Submitted',
     'form_mail_title' => 'Contact form',
     'form_mail_url' => config('app.url'),
@@ -17,30 +17,39 @@ return [
     'form_subject_received' => 'Contact form received',
     'form_subject_sent' => 'Contact form sent',
     'form_success' => 'Thank you, we received your message and will contact you as soon as possible.',
+    'form_required' => 'Required',
 
-    // Cookie banner
-    'cookie_accept_all' => 'Accept all',
-    'cookie_accept_selected' => 'Accept selected',
-    'cookie_ads' => 'Ads & tracking',
-    'cookie_ads_explanation' => 'Third party cookies used for advertisement.',
-    'cookie_analytics' => 'Analytics',
-    'cookie_analytics_explanation' => 'Third party cookies used for analytics.',
-    'cookie_explanation' => 'This site uses cookies that need consent.',
-    'cookie_embeds' => 'Video embeds',
-    'cookie_embeds_explanation' => 'Third party cookies used for embedded video\'s that can possibly track you.',
-    'cookie_embeds_disabled' => 'Accept video embed cookies to view this video.',
-    'cookie_functional' => 'Functional',
-    'cookie_functional_explanation' => 'First party cookies needed for the website to function. These can\'t be turned off.',
-    'cookie_title' => 'Cookies 🍪',
-    'cookie_learn_more' => 'Learn more',
-    'cookie_change_preferences' => 'Change cookie preferences',
+    // Consent banner
+    'consent_accept_all' => 'Accept all',
+    'consent_accept_selected' => 'Accept selected',
+    'consent_reject_all' => 'Reject all',
+    'consent_ads' => 'Ads & tracking',
+    'consent_ads_explanation' => 'Storage (like third party cookies) used for advertisement.',
+    'consent_ad_user_data' => 'User data',
+    'consent_ad_user_data_explanation' => 'Send user data related to advertising to Google.',
+    'consent_ad_personalization' => 'Ad personalization',
+    'consent_ad_personalization_explanation' => 'Get personalized advertising.',
+    'consent_analytics' => 'Analytics',
+    'consent_analytics_explanation' => 'Storage (like third party cookies) used for analytics.',
+    'consent_explanation' => 'This site uses third party services that need your consent.',
+    'consent_embeds' => 'Video embeds',
+    'consent_embeds_explanation' => 'Third party cookies used for embedded video\'s that can possibly track you.',
+    'consent_embeds_accept' => 'Give consent to third party embeds to view this content.',
+    'consent_embeds_open' => 'Open consent banner to accept embeds.',
+    'consent_functional' => 'Functional',
+    'consent_functional_explanation' => 'First party cookies needed for the website to function. These can\'t be turned off.',
+    'consent_title' => 'Consent',
+    'consent_learn_more' => 'Learn more',
+    'consent_change_preferences' => 'Manage consent',
+    'consent_more_info' => 'More information',
 
     // Footer
     'on' => 'on',
 
     // Navigation
     'close' => 'Close',
-    'main_navigation' => 'Main navigation',
+    'main_navigation_desktop' => 'Main navigation desktop',
+    'main_navigation_mobile' => 'Main navigation mobile',
     'menu' => 'Menu',
     'nav_close' => 'Close navigation',
     'nav_open' => 'Open navigation',
@@ -49,6 +58,9 @@ return [
     'skip_to_content' => 'Skip to content',
 
     // Other
+    'collection_mounted' => 'Your entry :title has a collection mounted and can\'t be deleted.',
+    'external_link' => '(external link)',
+    'home' => 'Homepage',
     'no_results' => 'There are currently no posts.',
     'no_script' => 'This website requires Javascript for some parts to function properly. Your experience may vary.',
 
@@ -56,25 +68,4 @@ return [
     'next' => 'next',
     'of' => 'of',
     'previous' => 'previous',
-
-    // Social images
-    'social_images' => 'Generated one social image.|Generated :count social images.',
-    'social_images_queue' => 'Generating one social image in the background.|Generating :count social images in the background.',
-
-    // Widgets
-    'widget_assets_title' => 'Assets without Alt Text in :container',
-    'widget_assets_edit' => 'Edit this Asset',
-    'widget_assets_explanation' => 'It\'s important to add alt text describing your images. This helps users who depend on assistive technology.',
-    'widget_assets_count' => '{0}You have :amount images that need attention.|{1}You have :amount image that needs attention.|[2,*]You have at least :amount images that need attention.',
-    'widget_assets_done' => 'All assets have an alt text.',
-
-    // News
-    'news_all' => 'All articles',
-    'news_more' => 'More articles',
-    'news_read_more' => 'Read more',
-
-    // Projects
-    'projects_all' => 'All articles',
-    'projects_more' => 'More articles',
-    'projects_read_more' => 'Read more',
 ];
