@@ -1,7 +1,27 @@
 <?php
 
 return [
-    // Contact form
+    'close' => 'Sluit',
+    'collection_mounted' => 'Je entry :title bevat een collectie en kan niet worden verwijderd.',
+    'consent_accept_all' => 'Accepteer alles',
+    'consent_accept_selected' => 'Accepteer selectie',
+    'consent_ads' => 'Advertenties & tracking',
+    'consent_ads_explanation' => 'Third party cookies benodigd voor het retargeten en plaatsen van advertenties.',
+    'consent_analytics' => 'Analytics',
+    'consent_analytics_explanation' => 'Third party cookies die nodig zijn voor het meten van geanomiseerde bezoekersstatistieken.',
+    'consent_change_preferences' => 'Manage toestemmingen',
+    'consent_embeds' => 'Video embeds',
+    'consent_embeds_accept' => 'Geef toestemming voor embeds van derden om deze inhoud te bekijken.',
+    'consent_embeds_open' => 'Open consent-banner om embeds te accepteren.',
+    'consent_embeds_explanation' => 'Third party cookies die nodig zijn om embedded video\'s te tonen. Deze kunnen je gedrag mogelijk tracken.',
+    'consent_explanation' => 'Deze site gebruikt cookies waar we je toestemming voor nodig hebben.',
+    'consent_functional' => 'Functioneel',
+    'consent_functional_explanation' => 'First party cookies die essentieel zijn voor het functioneren van deze website. Deze kunnen niet uitgeschakeld worden.',
+    'consent_learn_more' => 'Meer informatie',
+    'consent_more_info' => 'Meer informatie',
+    'consent_reject_all' => 'Weiger alles',
+    'consent_title' => 'Cookies 🍪',
+    'external_link' => '(externe link)',
     'form_error' => 'Er zijn problemen met je formulierinput. Probeer het opnieuw.',
     'form_honeypot' => 'Fax',
     'form_mail_body_owner' => 'Er is een formulier verzonden.',
@@ -9,7 +29,8 @@ return [
     'form_mail_closing' => 'Vriendelijke groeten',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Beste',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
+    'form_required' => 'Verplicht',
     'form_mail_submitted' => 'Verzonden',
     'form_mail_title' => 'Contactformulier',
     'form_mail_url' => config('app.url'),
@@ -17,54 +38,19 @@ return [
     'form_subject_received' => 'Contactformulier ontvangen',
     'form_subject_sent' => 'Contactformulier verzonden',
     'form_success' => 'Bedankt voor je bericht.',
-
-    // Cookie banner
-    'cookie_accept_all' => 'Accepteer alles',
-    'cookie_accept_selected' => 'Accepteer selectie',
-    'cookie_ads' => 'Advertenties & tracking',
-    'cookie_ads_explanation' => 'Third party cookies benodigd voor het retargeten en plaatsen van advertenties.',
-    'cookie_analytics' => 'Analytics',
-    'cookie_analytics_explanation' => 'Third party cookies die nodig zijn voor het meten van geanomiseerde bezoekersstatistieken.',
-    'cookie_explanation' => 'Deze site gebruikt cookies waar we je toestemming voor nodig hebben.',
-    'cookie_embeds' => 'Video embeds',
-    'cookie_embeds_explanation' => 'Third party cookies die nodig zijn om embedded video\'s te tonen. Deze kunnen je gedrag mogelijk tracken.',
-    'cookie_embeds_disabled' => 'Accepteer video-embed-cookies om deze video te zien.',
-    'cookie_functional' => 'Functioneel',
-    'cookie_functional_explanation' => 'First party cookies die essentieel zijn voor het functioneren van deze website. Deze kunnen niet uitgeschakeld worden.',
-    'cookie_title' => 'Cookies 🍪',
-    'cookie_learn_more' => 'Meer informatie',
-    'cookie_change_preferences' => 'Wijzig cookie-voorkeuren',
-
-    // Footer
-    'on' => 'op',
-
-    // Navigation
-    'close' => 'Sluit',
-    'main_navigation' => 'Hoofdnavigatie',
+    'home' => 'Startpagina',
+    'main_navigation_desktop' => 'Hoofdnavigatie desktop',
+    'main_navigation_mobile' => 'Hoofdnavigatie mobile',
     'menu' => 'Menu',
     'nav_close' => 'Sluit navigatie',
     'nav_open' => 'Open navigatie',
-    'subnav_close' => 'Sluit subnavigatie',
-    'subnav_open' => 'Open subnavigatie',
-    'skip_to_content' => 'Doorgaan naar content',
-
-    // Other
+    'next' => 'Volgende',
     'no_results' => 'Er zijn momenteel geen resultaten.',
     'no_script' => 'Deze website vereist Javascript om sommige onderdelen goed te laten functioneren.',
-
-    // Pagination
-    'next' => 'Volgende',
     'of' => 'van',
+    'on' => 'op',
     'previous' => 'Vorige',
-
-    // Social images
-    'social_images' => 'Generated one social image.|Generated :count social images.',
-    'social_images_queue' => 'Generating one social image in the background.|Generating :count social images in the background.',
-
-    // Widgets
-    'widget_assets_title' => 'Assets zonder alt tekst in :container',
-    'widget_assets_edit' => 'Wijzig deze asset',
-    'widget_assets_explanation' => 'Het is belangrijk om alt-teksten toe te voegen die je afbeeldingen omschrijven. Dit helpt gebruikers die afhankelijk zijn van ondersteunende technologie.',
-    'widget_assets_count' => '{0}|{1}Je hebt :amount afbeelding die je aandacht nodig heeft.|[2,*]Je hebt (tenminste) :amount afbeeldingen die je aandacht nodig hebben.',
-    'widget_assets_done' => 'Alle afbeeldingen hebben een alt-tekst.',
+    'skip_to_content' => 'Doorgaan naar content',
+    'subnav_close' => 'Sluit subnavigatie',
+    'subnav_open' => 'Open subnavigatie',
 ];
