@@ -8,7 +8,7 @@ return [
     'form_mail_closing' => 'Vennleg helsing',
     'form_mail_from' => config('app.name'),
     'form_mail_greeting' => 'Kjære',
-    'form_mail_logo' => 'https://cdn.studio1902.nl/assets/statamic-peak/statamic-peak-logo.png',
+    'form_mail_logo' => config('app.url').'/visuals/statamic-peak-logo.png',
     'form_mail_submitted' => 'Levert',
     'form_mail_title' => 'Kontaktskjema',
     'form_mail_url' => config('app.url'),
@@ -16,18 +16,19 @@ return [
     'form_success' => 'Tusen takk, vi har motteke meldinga di og kontaktar deg så snart som råd.',
 
     // Cookie banner
-    'cookie_title' => 'Informasjonskapslar 🍪',
-    'cookie_functional' => 'Fungerande',
-    'cookie_functional_explanation' => 'Førstepartsinformasjonskapslar som trengs for at nettsida skal fungere. Du kan ikkje slå desse av.',
-    'cookie_analytics' => 'Analysar',
-    'cookie_analytics_explanation' => 'Tredjepartsinformasjonskapslar, trengs ikkje.',
-    'cookie_ads' => 'Reklame og sporing',
-    'cookie_ads_explanation' => 'Tredjepartsinformasjonskapslar, trengs ikkje.',
-    'cookie_accept' => 'Aksepter',
-    'cookie_ignore' => 'Ignorer',
-    'cookie_explanation' => 'Denne nettstaden bruker informasjonskapslar som treng godkjenning.',
-    'cookie_learn_more' => 'Les meir',
-    'cookie_reset_consent' => 'Gjenopprett godkjenning av informasjonskapsel',
+    'consent_title' => 'Informasjonskapslar 🍪',
+    'consent_functional' => 'Fungerande',
+    'consent_functional_explanation' => 'Førstepartsinformasjonskapslar som trengs for at nettsida skal fungere. Du kan ikkje slå desse av.',
+    'consent_analytics' => 'Analysar',
+    'consent_analytics_explanation' => 'Tredjepartsinformasjonskapslar, trengs ikkje.',
+    'consent_reject_all' => 'Reject all',
+    'consent_ads' => 'Reklame og sporing',
+    'consent_ads_explanation' => 'Tredjepartsinformasjonskapslar, trengs ikkje.',
+    'consent_accept' => 'Aksepter',
+    'consent_ignore' => 'Ignorer',
+    'consent_explanation' => 'Denne nettstaden bruker informasjonskapslar som treng godkjenning.',
+    'consent_learn_more' => 'Les meir',
+    'consent_reset_consent' => 'Gjenopprett godkjenning av informasjonskapsel',
 
     // Footer
     'on' => 'på',
@@ -44,6 +45,7 @@ return [
     // Other
     'dark_mode_on' => 'Slå på mørk modus',
     'dark_mode_off' => 'Slå av mørk modus',
+    'home' => 'Homepage',
     'no_results' => 'Det er ingen innlegg for augneblinken.',
     'no_script' => 'Denne nettstaden treng Javascript for at alle deler skal fungere korrekt. Opplevinga di kan variere om du ikkje brukar Javascript.',
 
@@ -56,10 +58,6 @@ return [
     'search' => 'Søk',
     'search_no_results' => 'Fann ingen resultat',
     'search_results_for' => 'Resultat for',
-
-    // Social images
-    'social_images' => 'Genererte eit bilete for sosiale medium.|Genererte :count bilete for sosiale medium.',
-    'social_images_queue' => 'Genererte eit bilete for sosiale medium i bakgrunnen.|Genererte :count bilete for sosiale medium i bakgrunnen.',
 
     // Other
     'vacancy_visit' => 'Sjå stillingsannonse',
