@@ -1,32 +1,21 @@
 import Alpine from 'alpinejs'
 import collapse from '@alpinejs/collapse'
+import focus from '@alpinejs/focus'
 import morph from '@alpinejs/morph'
 import persist from '@alpinejs/persist'
-import focus from '@alpinejs/focus'
 import intersect from '@alpinejs/intersect'
-import 'focus-visible'
+import precognition from 'laravel-precognition-alpine';
 
 //Custom Modules
 import consoleHello from './modules/consoleHello';
 import preloader from './modules/preloader';
 import tinyslider from './modules/tinyslider';
 
-// Global get CSRF token function (used by forms).
-window.getToken = async () => {
-    return await fetch('/!/statamic-peak-tools/dynamic-token/refresh')
-        .then((res) => res.json())
-        .then((data) => {
-            return data.csrf_token
-        })
-        .catch((error) => {
-            this.error = 'Something went wrong. Please try again later.'
-        })
-}
-
 // Call Alpine.
 window.Alpine = Alpine
-Alpine.plugin([collapse, focus, morph, persist, intersect])
+Alpine.plugin([collapse, focus, morph, persist, precognition, intersect])
 Alpine.start()
+
 
 // Call custom modules.
 const CDG = {
@@ -35,7 +24,6 @@ const CDG = {
         //consoleHello('CDG is ready');
         //preloader();
         tinyslider();
-        
 
         window.addEventListener("resize", function(){
             //consoleHello('window has resized');
@@ -57,4 +45,4 @@ CDG.onreadyFunctions();
 
 window.onload = function(event) {
     CDG.onloadFunctions();
-}; 
+};
